requirements.txt
django-environ==0.4.5
django-htmlmin==0.11.0
Django==3.2.5
djangorestframework==3.12.4
gunicorn==20.1.0
holidays==0.11.1
pytz==2021.1